Rename components to views
