+++ src/views/landing/index.js
